Move StratosStatus
