+++ src/frontend/packages/cloud-foundry/src/features/cloud-foundry/services/cloud-foundry-org-space-quota.ts
@@ -3,9 +3,9 @@ import { filter, first, map, switchMap } from 'rxjs/operators';
 
 import { truthyIncludingZero } from '../../../../../core/src/core/utils.service';
 import { determineCardStatus } from '../../../../../core/src/shared/components/cards/card-status/card-status.component';
-import { StratosStatus } from '../../../../../core/src/shared/shared.types';
 import { EntityMonitorFactory } from '../../../../../store/src/monitors/entity-monitor.factory.service';
 import { APIResource } from '../../../../../store/src/types/api.types';
+import { StratosStatus } from '../../../../../store/src/types/shared.types';
 import { IApp, IOrganization, ISpace } from '../../../cf-api.types';
 import { CFEntityConfig } from '../../../cf-types';
 import { CloudFoundryEndpointService } from './cloud-foundry-endpoint.service';
