+++ src/frontend/packages/cloud-foundry/src/features/cloud-foundry/services/cloud-foundry-space-quota.ts
@@ -1,9 +1,9 @@
 import { Observable } from 'rxjs';
 
 import { spaceEntityType } from '../../../../../cloud-foundry/src/cf-entity-types';
-import { StratosStatus } from '../../../../../core/src/shared/shared.types';
 import { EntityMonitorFactory } from '../../../../../store/src/monitors/entity-monitor.factory.service';
 import { APIResource } from '../../../../../store/src/types/api.types';
+import { StratosStatus } from '../../../../../store/src/types/shared.types';
 import { IApp, ISpace } from '../../../cf-api.types';
 import { getStartedAppInstanceCount } from '../../../cf.helpers';
 import { CloudFoundryEndpointService } from './cloud-foundry-endpoint.service';
