+++ src/frontend/packages/cloud-foundry/src/shared/components/service-plan-public/service-plan-public.component.ts
@@ -6,8 +6,8 @@ import {
   getServicePlanAccessibilityCardStatus,
 } from '../../../../../cloud-foundry/src/features/service-catalog/services-helper';
 import { ServicesService } from '../../../../../cloud-foundry/src/features/service-catalog/services.service';
-import { StratosStatus } from '../../../../../core/src/shared/shared.types';
 import { APIResource } from '../../../../../store/src/types/api.types';
+import { StratosStatus } from '../../../../../store/src/types/shared.types';
 import { IServiceBroker, IServicePlan } from '../../../cf-api-svc.types';
 import { cfEntityCatalog } from '../../../cf-entity-catalog';
