+++ src/frontend/packages/core/src/core/stateful-icon/stateful-icon.component.spec.ts
@@ -1,6 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { StratosStatus } from '../../shared/shared.types';
+import { StratosStatus } from '../../../../store/src/types/shared.types';
 import { MDAppModule } from '../md.module';
 import { StatefulIconComponent } from './stateful-icon.component';
