+++ src/frontend/packages/core/src/features/error-page/error-page/error-page.component.ts
@@ -14,7 +14,7 @@ import { InternalEventMonitorFactory } from '../../../../../store/src/monitors/i
 import { EndpointModel } from '../../../../../store/src/types/endpoint.types';
 import { InternalEventState } from '../../../../../store/src/types/internal-events.types';
 import { getPreviousRoutingState } from '../../../../../store/src/types/routing.type';
-import { StratosStatus } from '../../../shared/shared.types';
+import { StratosStatus } from '../../../../../store/src/types/shared.types';
 import { eventReturnUrlParam } from '../../event-page/events-page/events-page.component';
 
 @Component({
