+++ src/frontend/packages/core/src/shared/components/application-state/application-state-icon/application-state-icon.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 
-import { StratosStatus } from '../../../shared.types';
+import { StratosStatus } from '../../../../../../store/src/types/shared.types';
 
 
 @Component({
