+++ src/frontend/packages/core/src/shared/components/cards/card-status/card-status.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
 import { Observable } from 'rxjs';
 
-import { StratosStatus } from '../../../shared.types';
+import { StratosStatus } from '../../../../../../store/src/types/shared.types';
 
 
 export function determineCardStatus(value: number, limit: number): StratosStatus {
