+++ src/frontend/packages/core/src/shared/components/favorites-meta-card/favorites-meta-card.component.ts
@@ -11,10 +11,10 @@ import { userFavoritesEntitySchema } from '../../../../../store/src/base-entity-
 import { entityCatalog } from '../../../../../store/src/entity-catalog/entity-catalog';
 import { IFavoritesMetaCardConfig } from '../../../../../store/src/favorite-config-mapper';
 import { endpointEntitiesSelector } from '../../../../../store/src/selectors/endpoint.selectors';
+import { ComponentEntityMonitorConfig, StratosStatus } from '../../../../../store/src/types/shared.types';
 import { IFavoriteMetadata, UserFavorite } from '../../../../../store/src/types/user-favorites.types';
 import { IFavoriteEntity } from '../../../../../store/src/user-favorite-manager';
 import { isEndpointConnected } from '../../../features/endpoints/connect.service';
-import { ComponentEntityMonitorConfig, StratosStatus } from '../../shared.types';
 import { ConfirmationDialogConfig } from '../confirmation-dialog.config';
 import { ConfirmationDialogService } from '../confirmation-dialog.service';
 import { MetaCardMenuItem } from '../list/list-cards/meta-card/meta-card-base/meta-card.component';
