+++ src/frontend/packages/core/src/shared/global-events.service.ts
@@ -3,8 +3,8 @@ import { Store } from '@ngrx/store';
 import { BehaviorSubject, combineLatest, Observable, ReplaySubject } from 'rxjs';
 import { debounceTime, distinctUntilChanged, map, publishReplay, refCount, startWith } from 'rxjs/operators';
 
+import { StratosStatus } from '../../../store/src/types/shared.types';
 import { GeneralEntityAppState } from './../../../store/src/app-state';
-import { StratosStatus } from './shared.types';
 
 export type GlobalEventTypes = 'warning' | 'error' | 'process' | 'complete';
